+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Breaking
 
 ### Features
+* add support for fractional UNIX timestamps in the `timestamper` processor while preserving integer timestamp normalization
 
 ### Improvements